backmaster/requirements.txt
absl-py==0.10.0
appdirs==1.4.4
astroid==2.4.2
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
cycler==0.10.0
Cython==0.29.21
defusedxml==0.6.0
distlib==0.3.1
Django==2.1.15
django-allauth==0.42.0
django-appconf==1.0.4
django-cors-headers==3.5.0
django-imagekit==4.0.2
django-rest-auth==0.9.5
djangorestframework==3.11.1
filelock==3.0.12
google-auth==1.21.3
google-auth-oauthlib==0.4.1
grpcio==1.32.0
idna==2.10
importlib-metadata==1.7.0
importlib-resources==3.0.0
isort==4.3.21
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
Markdown==3.2.2
matplotlib==3.0.3
mccabe==0.6.1
numpy==1.18.5
oauthlib==3.1.0
opencv-python==4.4.0.42
pilkit==2.0
Pillow==7.2.0
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pylint==2.5.3
pyparsing==2.4.7
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
scipy==1.4.1
six==1.15.0
sqlparse==0.3.1
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
toml==0.10.1
tqdm==4.50.0
typed-ast==1.4.1
urllib3==1.25.10
virtualenv==20.0.31
Werkzeug==1.0.1
wrapt==1.12.1
zipp==1.2.0
